upgrade import

+++ components/load.js
@@ -26,9 +26,9 @@ const importCreds = {
 };
 
 
-export default async function main(opts = { sendEvents: true, sendProfiles: true, sendAnnotations: false }) {
+export default async function main(opts = { sendEvents: true, sendProfiles: true, sendAnnotations: false, deleteStuff: false }) {
 	const directory = "output"; // TEMP_DIR + '/output';
-	const { sendEvents = true, sendProfiles = true, sendAnnotations = false } = opts;
+	const { sendEvents = true, sendProfiles = true, sendAnnotations = false, deleteStuff = false } = opts;
 	const fileSystem = (await ls(path.resolve(TEMP_DIR, directory)))
 		.filter((dir) => isDirOrFile(dir) === 'directory')
 		.map((dir) => details(dir))
@@ -40,7 +40,7 @@ export default async function main(opts = { sendEvents: true, sendProfiles: true
 
 
 	const results = { fileSystem };
-	if (sendProfiles) {
+	if (sendProfiles && deleteStuff) {
 		try {
 			if (NODE_ENV === 'dev') console.log(`\nDeleting profiles from Mixpanel\n`);
 			const profileDeletes = await deleteProfiles();
@@ -212,7 +212,7 @@ if (import.meta.url === new URL(`file://${process.argv[1]}`).href) {
 	// }
 	let result;
 	try {
-		result = await main({ sendEvents: false, sendProfiles: true });
+		result = await main({ sendEvents: true, sendProfiles: true, deleteStuff: false });
 		// result = await deleteProfiles();
 		// result = await loadChartAnnotations();
 	}
